Adapt program options for cluster.py
RAMCloud's cluster.py script requires clients to accept a small set
of standard options, including clientIndex and numClients. This
update adds those options, and for the TwitterWorkloadClient uses
these to replace serverNumber, baseServerNumber, and numServers,
which more amenable to ansible when we used ansible for launching
cluster jobs instead of cluster.py.